Crate conch

Conch - pretty shell

conch is a collection of structs, enums and implementations to make prettifying shell much easier.

Macros

ansi_enum_builder

Structs

ANSIEscapeCode

A basic dataclass of a deconstructed \x1b[00;00;..m structure.

BackgroundIter

An iterator over the variants of Self

ColourIter

An iterator over the variants of Self

ESCAPE_CODE_PATTERN

ESCAPE_CODE_START_PATTERN

IntensityIter

An iterator over the variants of Self

SEP_PATTERN

Enums

Background

An Enum for ANSI 256-colour codes.

Colour

An Enum for ANSI 256-colour codes.

Intensity

Intensity modifier

Modifier

ModifierError

A enum of possible error types during modifier operations.

Constants

DEFAULT_SEPARATOR

Default sep to use for ANSIEscapeCode.

Traits

IntoANSIEscapeCode

Custom Trait for anything that can be converted into ANSIEscapeCode.

Resetter

Trait for anything that has a resetter method.

StringWrapper

Auto trait for structs and enums that can wrap a string.