Struct conch::ANSIEscapeCode

pub struct ANSIEscapeCode {
    pub code: Option<u16>,
    pub modifiers: Vec<i32, Global>,
    pub sep: char,
    pub end_char: char,
    pub source_str: Option<String>,
}

A basic dataclass of a deconstructed \x1b[00;00;..m structure.

This dataclass simply represents a syntax valid ANSI Escape pattern; it does not necessarily guarantee that the pattern is meaningful or legal for the specific command.

Example

use conch_base_models::{ANSIEscapeCode, DEFAULT_SEPARATOR};

let parsed: ANSIEscapeCode = "\x1b[30m".try_into().unwrap();
assert_eq!(
    parsed,
    ANSIEscapeCode {
        code: Some(30),
        modifiers: Vec::new(),
        sep: DEFAULT_SEPARATOR,
        end_char: 'm',
        source_str: None, // This is not used in comparison
    }
);

let parsed: ANSIEscapeCode = "\x1b[20;8H".try_into().unwrap();
assert_eq!(
    parsed,
    ANSIEscapeCode {
        code: None,
        modifiers: vec![20,8],
        sep: DEFAULT_SEPARATOR,
        end_char: 'H',
        source_str: None, // This is not used in comparison
    }
);

let parsed: ANSIEscapeCode = "\x1b[38:5:255m".try_into().unwrap();
assert_eq!(
    parsed,
    ANSIEscapeCode {
        code: Some(38),
        modifiers: vec![5,255],
        sep: DEFAULT_SEPARATOR,
        end_char: 'm',
        source_str: None, // This is not used in comparison
    }
);

// Code is less than 0
let parsed: Result<ANSIEscapeCode, _> = "\x1b[-1m".try_into();
assert!(
    parsed.is_err(),
);

// End character missing
let parsed: Result<ANSIEscapeCode, _> = "\x1b[38;5;0".try_into();
assert!(
    parsed.is_err(),
);

// Pattern does not begin with `\x1b`
let parsed: Result<ANSIEscapeCode, _> = "FooBar\x1b[38;5;0".try_into();
assert!(
    parsed.is_err(),
);

Fields
code: Option<u16>

Command Code.

If end_char is 'm', then this will be a Some<u16> containing the first code after \x1b[. For example, for the colour code of \x1b[30m, code will be [Some(30_u16)]. If end_char is not 'm', code shall be None.

modifiers: Vec<i32, Global>

Modifier codes.

a Vec of i32 collecting all the modifier codes that follows code above. For commands like \x1b[2A to move cursor up by 2 rows, modifiers will be vec![2]; for colour commands like the above \x1b[30m, this will be an empty vec![].

sep: char

Separator char.

Must be either : or ; to be valid.

This is not currently in use when parsing; any code that is parsed will use DEFAULT_SEPARATOR instead; however if this is set, then to_string will build the String wtih the separator.

end_char: char

The trailing character of the sequence.

Typically 'm' for most modern commands; but can be any char within [A-Za-z] for other codes such as cursor movement.

Mandatory - without this char, the pattern cannot be terminated.

source_str: Option<String>

Original String that generated this instance.

This field is for the sole purpose of allowing len() to work. Instead of measuring the length of the rebuilt string, this field records the original String from TryFrom<&str>, so this value will always be accurate.

Implementations

impl ANSIEscapeCode

pub fn new(
code: Option<u16>,
modifiers: Option<Vec<i32, Global>>,
end_char: char
) -> ANSIEscapeCode

Creates a new ANSIEscapeCode instance with the default separator.

pub fn add_source(self, text: &str) -> ANSIEscapeCode

Chained method to add a soruce to this instance.

pub fn use_sep(self, sep: char) -> ANSIEscapeCode

Replace the seperator in this instance.

pub fn parse(text: &str) -> Result<Captures<'_>, ModifierError>

Parse a str beginning with \x1b into a [regex::Captures] object.