Struct conch::CalendarMonth

pub struct CalendarMonth<Region>where
    Region: RegionMarker,{
    pub date: NaiveDate,
    pub week_starts_with: Weekday,
    pub show_title: bool,
    pub show_other_months: bool,
    pub capitalize_title: bool,
    pub title_modifier: Modifier,
    pub other_month_modifier: Modifier,
    pub weekday_modifier: Modifier,
    pub holiday_modifier: Modifier,
    pub today_modifier: Option<Modifier>,
    pub decorated_days: HashMap<NaiveDate, Modifier, RandomState>,
    /* private fields */
}

A struct to display a month on a calendar in stdout.

Allows special highlighting to

public holidays,

days outside of the current month, or

any arbitrary days.

Example:

use chrono::NaiveDate;

use conch::{
    s, // macro for combining string literals, for assertion only.
    Modifier, // ANSI styling tool
    Lines, // Helper struct to format a [`Vec<String>`] for stdout.
    CalendarMonth,
    regions // Defining the region of [`CalendarMonth`].
};

let today_highlight =
    Modifier::colour("BrightBlue").unwrap()
    + Modifier::intensity("Bold").unwrap()
;
let today = NaiveDate::from_ymd_opt(2023, 3, 3).unwrap();

let calendar: CalendarMonth<regions::England> =
    CalendarMonth::new(today)
    .modify_today(Some(today_highlight))
;

// Works with both owned and `&calendar`
let lines = Lines::from(calendar);

println!("{}", &lines);
assert_eq!(lines.to_string(), s!(
    "\u{1b}[1m M  T  W  T  F  S  S\u{1b}[22m\n"
    "       1  2 \u{1b}[38;5;12m\u{1b}[1m 3\u{1b}[22m\u{1b}[39m  4 \u{1b}[38;5;9m\u{1b}[1m 5\u{1b}[22m\u{1b}[39m\n"
    " 6  7  8  9 10 11 \u{1b}[38;5;9m\u{1b}[1m12\u{1b}[22m\u{1b}[39m\n"
    "13 14 15 16 17 18 \u{1b}[38;5;9m\u{1b}[1m19\u{1b}[22m\u{1b}[39m\n"
    "20 21 22 23 24 25 \u{1b}[38;5;9m\u{1b}[1m26\u{1b}[22m\u{1b}[39m\n"
    "27 28 29 30 31      "
));

This will produce a printout like:

 M  T  W  T  F  S  S
       1  2 *3* 4 *5*
 6  7  8  9 10 11*12*
13 14 15 16 17 18*19*
20 21 22 23 24 25*26*
27 28 29 30 31

with *\d+* denoting a colour.

Assuming today is the 3rd, then 3 will be in bright blue as dictated by today_highlight above.

A number of chained modifiers are available. For example, the above example is equivalent to:

use chrono::{NaiveDate, Weekday};

use conch::{
    Modifier, // ANSI styling tool
    CalendarMonth,
    regions // Defining the region of [`CalendarMonth`].
};

let today_highlight =
    Modifier::colour("BrightBlue").unwrap()
    + Modifier::intensity("Bold").unwrap()
;
let today = NaiveDate::from_ymd_opt(2023, 3, 3).unwrap();

let calendar: CalendarMonth<regions::England> =
    CalendarMonth::new(today)
    .starts_week_with(Weekday::Mon)
    .show_title(true)
    .show_other_months(false)
    .capitalize_title(true)
    .modify_title(
        Modifier::colour("Grayscale13").unwrap()
    )
    .modify_weekdays(Modifier::Nothing)
    .modify_holidays(
        Modifier::colour("BrightRed").unwrap()
        + Modifier::intensity("Bold").unwrap()
    )
    .modify_today(Some(today_highlight))
;

Fields
date: NaiveDate

NaiveDate that this calendar is based on. The day will be discarded upon instantiation, and replaced with the first day of the month.

week_starts_with: Weekday

The Weekday to starts each week with. Each region can set its own default which is used unless overridden in this CalendarMonth instance.

Use Self::starts_week_with() to change this.

show_title: bool

Defines whether the weekday initials like M T W T F S S should be shown.

Use Self::show_title() to change this.

show_other_months: bool

Defines whether to show the days from another month, if the starting and ending weeks of the month extends into another month.

Use Self::show_other_months() to change this.

capitalize_title: bool

Defines whether to capitalize the title row, so that instead of M T W T F S S, m t w t f s s will be displayed.

Use Self::capitalize_title() to change this.

title_modifier: Modifier

Modifier for the title row.

Use Self::modify_title() to change this.

other_month_modifier: Modifier

Modifier for the days from other months. This will only be applied if Self::show_other_months is true.

Use Self::modify_other_months() to change this.

weekday_modifier: Modifier

Modifier for any undecorated weekdays within the month.

Use Self::modify_weekdays() to change this.

holiday_modifier: Modifier

Modifier for any Sundays and Public holidays within the month.

Use Self::modify_holidays() to change this.

today_modifier: Option<Modifier>

Modifier for today, if present within the month. This differs from other modifiers in that an Option<Modifier> is expected instead; this is due to the fact that if this is None, then Self::weekday_modifier or Self::holiday_modifier will be used instead, whichever relevant.

Use Self::modify_today() to change this.

decorated_days: HashMap<NaiveDate, Modifier, RandomState>

A hashmap of days that requires special modifiers.

Each calendar can register special arbitrary days within itself, and use a special modifier on each of them. This allows you to customise the calendar to the highlighting needs of your app.

Implementations

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn new(month: NaiveDate) -> CalendarMonth<Region>

Create a new CalendarMonth from a NaiveDate provided.

pub fn decorate_day(
self,
date: NaiveDate,
modifier: Modifier
) -> CalendarMonth<Region>

Add a special Modifier to a single date.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn modify_title(self, value: Modifier) -> CalendarMonth<Region>

Chained method to change the title_modifier of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn modify_weekdays(self, value: Modifier) -> CalendarMonth<Region>

Chained method to change the weekday_modifier of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn modify_holidays(self, value: Modifier) -> CalendarMonth<Region>

Chained method to change the holiday_modifier of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn modify_today(self, value: Option<Modifier>) -> CalendarMonth<Region>

Chained method to change the today_modifier of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn modify_other_months(self, value: Modifier) -> CalendarMonth<Region>

Chained method to change the other_month_modifier of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn starts_week_with(self, value: Weekday) -> CalendarMonth<Region>

Chained method to change the week_starts_with of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn capitalize_title(self, value: bool) -> CalendarMonth<Region>

Chained method to change the capitalize_title of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn show_title(self, value: bool) -> CalendarMonth<Region>

Chained method to change the show_title of a CalendarMonth instance.

impl<Region> CalendarMonth<Region>where
Region: RegionMarker,

pub fn show_other_months(self, value: bool) -> CalendarMonth<Region>

Chained method to change the show_other_months of a CalendarMonth instance.