Trait conch::FindModifiers

pub trait FindModifiers {
    fn iter_modifiers(&self) -> ModifiersInText<'_, '_> ;
    fn len(&self) -> usize;
}

Trait for str and String to iterate its modifiers.

len is just wrapper around its respective len functions, to provide a guarantee that anything that impl FindModifiers will have a len function.

Required Methods

fn iter_modifiers(&self) -> ModifiersInText<'_, '_>

fn len(&self) -> usize