Trait conch::LastWeekdayOfMonth

pub trait LastWeekdayOfMonth {
    fn last_weekday_of_month_opt(
        year: i32,
        month: u32,
        weekday: Weekday
    ) -> Option<NaiveDate>;

    fn last_weekday_of_month(year: i32, month: u32, weekday: Weekday) -> NaiveDate { ... }
}

Trait for adding last_weekday_of_month_opt method to NaiveDate.

Required Methods

fn last_weekday_of_month_opt(
year: i32,
month: u32,
weekday: Weekday
) -> Option<NaiveDate>

Get the last weekday of the specified month.

As there always is a last weekday of a month, this method is guaranteed to succeed despite returning an Option<NaiveDate>; this is to match the syntax of NaiveDate::from_weekday_of_month_opt().

Provided Methods

fn last_weekday_of_month(year: i32, month: u32, weekday: Weekday) -> NaiveDate

Get the last weekday of the specified month.

As there always is a last weekday of a month, this method is guaranteed to succeed.